Module Expert.Focus

module By_row : sig ... end

module By_cell : sig ... end

type ('a, 'p, 'k, 'c) t =

| None : (unit, unit, 'k, 'c) t

| By_row : {

on_change : ('k option -> unit Bonsai_web.Effect.t) Bonsai_web.Bonsai.t;
(*
Row-selection is not required to be inside the viewport, so the selected row can be offscreen such that it isn't given to the table component. compute_presence forces the user to consider if a row is considered 'focused' or not.
*)

compute_presence : 'k option Bonsai_web.Bonsai.t -> Bonsai_web.Bonsai.graph @ local -> 'p Bonsai_web.Bonsai.t;
(*
A user might try to focus-by-key a row that has not been filtered out, but is not inside the viewport. In that case, key_rank will be used as a fallback to compute the desired index. If the effect returns `None`, the key does not correspond to a row under the current filter conditions, and the focus will be a no-op.
*)

key_rank : ('k -> int option Bonsai_web.Effect.t) Bonsai_web.Bonsai.t;

} -> (('k, 'p) Focus_by_row.t, 'p, 'k, 'c) t

| By_cell : {

on_change : (('k * 'c) option -> unit Bonsai_web.Effect.t) Bonsai_web.Bonsai.t;

compute_presence : ('k * 'c) option Bonsai_web.Bonsai.t -> Bonsai_web.Bonsai.graph @ local -> 'presence Bonsai_web.Bonsai.t;

key_rank : ('k -> int option Bonsai_web.Effect.t) Bonsai_web.Bonsai.t;

} -> (('k, 'c, 'presence) By_cell.t, 'presence, 'k, 'c) t